Changed the basic type for disc thresholds to Float. Continuous variables ordered for sparse data.

+++ src/main/scala/org/lidiagroup/hmourit/tfg/discretization/DiscretizerModel.scala
@@ -17,6 +17,6 @@ trait DiscretizerModel[T] extends Serializable {
    */
   def discretize(data: RDD[T]): RDD[T]
   
-  def getThresholds: Array[(Int, Seq[Double])]
+  def getThresholds: Array[(Int, Seq[Float])]
 
 }
